tccl: tests fixes
      Common init helpers + change to new lib api
